requirements.txt
Django
python-dotenv
termcolor
validate-docbr
pytest
pytest-cov
psycopg2
psycopg2-binary
pylint-django
djangorestframework
markdown
django-filter
pyyaml
uritemplate
pygments
django-guardian

Faker
pytz
six
sqlparse

django-unfold
django-cors-headers

django-redis
djangorestframework-xml
djangorestframework-yaml

#django-admin-honeypot
git+https://github.com/dmpayton/django-admin-honeypot.git@develop

drf-yasg

## Unknow
# python-dateutils
# text-unicode

# Dev tools
coverage
cx_Freeze
flake8
icecream
mypy
pylint
setuptools-version-command
